pyproject.toml
[project]
name = "zlog-discord"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Yazdani",email = "yazdmich@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "sqlacodegen-v2 (>=0.1.4,<0.2.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "sqlacodegen-v2 (>=0.1.4,<0.2.0)"
]
